Remove unwanted urls and fix delete

+++ src/game/views.py
@@ -1,6 +1,7 @@
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
+from django.contrib import messages
 from django.views import generic
 from django.db.models import Count
 from participants.forms import ModifyUserForm
@@ -19,10 +20,6 @@ class DetailView(generic.DetailView) :
     model = Player
     template_name = 'game/detail.html'
 
-class PointView(generic.DetailView) :
-    model = Player
-    template_name = 'game/points.html'
-
 def detail(request, player_id) :
     player = get_object_or_404(Player, pk=player_id)
 
@@ -31,28 +28,18 @@ def detail(request, player_id) :
 
     return render(request, 'game/detail.html', {'player' : player, 'register_form': register_form, 'player_form': player_form})
 
-
-def attribute(request, player_id):
-    player = get_object_or_404(Player, pk=player_id)
-    try :
-        if request.POST['date'] == "" :
-            action = Action(player = player, point = request.POST['point'])
-        else :
-            action = Action(player = player, point = request.POST['point'], act_date = request.POST['date'])
-    except (KeyError, Player.DoesNotExist) :
-        return render(request, 'game/detail.html', {
-            'player': player,
-            'error_message': "Les entrées ne sont pas valides",
-        })
-    else :
-        action.save()
-        return HttpResponseRedirect(reverse('game:points', args=(player.id,)))
-
 def delete(request, player_id) :
-    player = get_object_or_404(Player, pk=player_id)
-    #player = Player.objects.get(pk=player_id)
-    player.delete()
-    return HttpResponseRedirect(reverse('game:index'))
+    if request.user.is_authenticated :
+        if request.user.id == player_id :
+            player = get_object_or_404(Player, pk=player_id)
+            player.delete()
+            return HttpResponseRedirect(reverse('game:index'))
+        else :
+            messages.error(request, "Tu ne peux pas supprimer un autre joueur que toi petit malin")
+            return HttpResponseRedirect(reverse('game:index'))
+    else :
+        messages.error(request, "Il faut être connecté pour pouvoir supprimer un joueur")
+        return HttpResponseRedirect(reverse('game:index'))
 
 def add_action(request) :
     submitted = False
